back/requirements.txt
asgiref==3.7.2
autopep8==2.0.4
blinker==1.7.0
Brotli==1.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
ConfigArgParse==1.7
contourpy==1.2.0
cryptography==41.0.5
cycler==0.12.1
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==4.2.7
django-allauth==0.58.2
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3
django-seed==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Faker==20.1.0
Flask==3.0.0
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.0
fonttools==4.45.0
gevent==23.9.1
geventhttpclient==2.0.11
greenlet==3.0.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
locust==2.19.0
MarkupSafe==2.1.3
matplotlib==3.8.2
msgpack==1.0.7
numpy==1.26.2
oauthlib==3.2.2
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
psutil==5.9.6
pycodestyle==2.11.1
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
pywin32==306
pyzmq==25.1.1
requests==2.31.0
requests-oauthlib==1.3.1
roundrobin==0.0.4
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.2.0
tomli==2.0.1
toposort==1.10
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
Werkzeug==3.0.1
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
